refactor(frontend): shared map components
`BaseMap` and `DataMap` both depend on modules outside the `lib` directory now, so move them to the `map` directory.
