+++ frontend/src/map/DataMap.tsx
@@ -10,11 +10,11 @@ import { backgroundState, showLabelsState } from 'map/layers/layers-state';
 import { useBasemapStyle } from 'map/use-basemap-style';
 import { useInteractions } from 'lib/state/interactions/use-interactions';
 
-import { useTriggerMemo } from '../hooks/use-trigger-memo';
-import { useDataLoadTrigger } from './use-data-load-trigger';
+import { useTriggerMemo } from 'lib/hooks/use-trigger-memo';
+import { useDataLoadTrigger } from 'lib/data-map/use-data-load-trigger';
 
-import { DeckGLOverlay } from '../map/DeckGLOverlay';
-import { ViewLayer, ViewLayerParams } from './view-layers';
+import { DeckGLOverlay } from 'lib/map/DeckGLOverlay';
+import { ViewLayer, ViewLayerParams } from 'lib/data-map/view-layers';
 import { LayersList } from 'deck.gl/typed';
 
 // set a convention where the view layer id is either the first part of the deck id before the @ sign, or it's the whole id
